# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Jandaira API
  description: |
    REST API for Jandaira — an AI agent swarm orchestration system.

    Most endpoints require the hive to be configured first via `POST /api/setup`.
    Unconfigured requests return `428 Precondition Required`.

    Real-time progress (swarm status, tool calls, approvals) is streamed over WebSocket at `GET /ws`.
  version: 1.0.0
  contact:
    name: Francisco Damião Terto
servers:
  - url: http://localhost:{port}
    description: Local server
    variables:
      port:
        default: "8080"
...
paths:
  /api/setup:
    post:
      tags: [Setup]
      summary: Configure the hive
      description: |
        Initializes the hive on first run. Sets the default swarm name, LLM model,
        token budget, and optionally stores the OpenAI API key in the local vault.

        Once configured this endpoint returns `403 Forbidden`.
        Does **not** require prior configuration (exempt from setup middleware).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupRequest'
            examples:
              minimal:
                summary: Minimal setup
                value:
                  api_key: sk-...
              full:
                summary: Full setup
                value:
                  api_key: sk-...
                  model: gpt-4o
                  swarm_name: my-swarm
                  max_nectar: 40000
                  max_agents: 5
                  supervised: true
                  isolated: false
      responses:
        '200':
          description: Hive configured successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Configuração salva com sucesso! O ecossistema está pronto e as operárias acordaram."
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: Hive already configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/dispatch:
    post:
      tags: [Dispatch]
      summary: Dispatch a stateless swarm mission
      description: |
        Assembles a dynamic agent pipeline for the given goal and runs the
        workflow asynchronously (timeout: 10 min). The HTTP response returns
        immediately with `202 Accepted`; all progress events are pushed to
        connected WebSocket clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchRequest'
            example:
              goal: "Summarize the latest news about AI and write a 5-paragraph report"
              group_id: "my-swarm"
      responses:
        '202':
          description: Mission accepted and dispatched asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Mission dispatched to the swarm. Follow progress via WebSocket."
        '400':
          $ref: '#/components/responses/BadRequest'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/tools:
    get:
      tags: [Catalog]
      summary: List registered tools
      description: Returns all tools currently registered in the Queen with their name, description, and parameter schema.
      responses:
        '200':
          description: Tool list
          content:
            application/json:
              schema:
                type: object
                properties:
                  tools:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tool'
        '428':
          $ref: '#/components/responses/NotConfigured'
  /api/agents:
    get:
      tags: [Catalog]
      summary: List registered agents
      description: Returns all agents registered in the Queen. Currently returns an empty array (reserved for future use).
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      type: object
        '428':
          $ref: '#/components/responses/NotConfigured'
  /api/sessions:
    get:
      tags: [Sessions]
      summary: List all sessions
      description: Returns all sessions ordered by creation date (newest first). Agent details are omitted — use `GET /api/sessions/{id}` for the full graph.
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
                  total:
                    type: integer
                    example: 3
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Sessions]
      summary: Create a session
      description: |
        Creates a new session with `active` status. After creation, dispatch the
        swarm via `POST /api/sessions/{id}/dispatch`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              name: "Weekly report"
              goal: "Generate the weekly performance report for Q2"
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  session:
                    $ref: '#/components/schemas/Session'
              example:
                message: "Sessão criada. Use POST /api/sessions/:id/dispatch para iniciar o enxame."
                session:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "Weekly report"
                  goal: "Generate the weekly performance report for Q2"
                  status: "active"
                  created_at: "2026-04-11T10:00:00Z"
                  updated_at: "2026-04-11T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/sessions/{id}:
    parameters:
      - $ref: '#/components/parameters/SessionID'

    get:
      tags: [Sessions]
      summary: Get a session
      description: Returns the full session object including all associated agents.
      responses:
        '200':
          description: Session with agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionWithAgents'
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Sessions]
      summary: Delete a session
      description: Permanently removes the session and all its agents (cascade delete).
      responses:
        '200':
          description: Session deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Sessão e agentes removidos com sucesso."
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/sessions/{id}/dispatch:
    post:
      tags: [Sessions]
      summary: Dispatch swarm for a session
      description: |
        Assembles the agent pipeline for the session, persists each specialist as an
        `Agent` record, and runs the workflow asynchronously (timeout: 10 min).

        Returns `409 Conflict` if the session is already in `completed` or `failed` status.

        The `goal` in the request body overrides the session's original goal for this run.
        All progress events are pushed to connected WebSocket clients.
      parameters:
        - $ref: '#/components/parameters/SessionID'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                goal:
                  type: string
                  description: Overrides the session goal for this dispatch. Falls back to the session's stored goal if omitted.
                  example: "Focus only on the revenue section of the report"
      responses:
        '202':
          description: Mission dispatched asynchronously
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  session_id:
                    type: string
                    format: uuid
                  agents:
                    type: integer
                    description: Number of specialists assembled
              example:
                message: "Missão despachada. Acompanhe o progresso via WebSocket."
                session_id: "550e8400-e29b-41d4-a716-446655440000"
                agents: 3
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Session already finished
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Sessão já finalizada com status 'completed'. Crie uma nova sessão."
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/sessions/{id}/agents:
    get:
      tags: [Sessions]
      summary: List session agents
      description: Returns all agents belonging to the session.
      parameters:
        - $ref: '#/components/parameters/SessionID'
      responses:
        '200':
          description: Agent list for the session
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  total:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/colmeias:
    get:
      tags: [Colmeias]
      summary: List all hives
      responses:
        '200':
          description: Hive list
          content:
            application/json:
              schema:
                type: object
                properties:
                  colmeias:
                    type: array
                    items:
                      $ref: '#/components/schemas/Colmeia'
                  total:
                    type: integer
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Colmeias]
      summary: Create a hive
      description: |
        Creates a persistent named hive. Set `queen_managed: false` to use
        pre-defined agents; `true` (default) lets the Queen assemble agents
        automatically on every dispatch.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateColmeiaRequest'
            examples:
              queen_managed:
                summary: Queen-managed hive
                value:
                  name: "Research Hive"
                  description: "General-purpose research"
                  queen_managed: true
              user_defined:
                summary: User-defined agents
                value:
                  name: "Code Review Hive"
                  description: "Fixed pipeline for code reviews"
                  queen_managed: false
      responses:
        '201':
          description: Hive created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  colmeia:
                    $ref: '#/components/schemas/Colmeia'
        '400':
          $ref: '#/components/responses/BadRequest'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/colmeias/{id}:
    parameters:
      - $ref: '#/components/parameters/ColmeiaID'

    get:
      tags: [Colmeias]
      summary: Get a hive with its agents
      responses:
        '200':
          description: Hive with agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColmeiaWithAgentes'
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Colmeias]
      summary: Update a hive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateColmeiaRequest'
      responses:
        '200':
          description: Hive updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  colmeia:
                    $ref: '#/components/schemas/Colmeia'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Colmeias]
      summary: Delete a hive
      description: Permanently removes the hive, all its agents, and its dispatch history (cascade delete).
      responses:
        '200':
          description: Hive deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/colmeias/{id}/dispatch:
    post:
      tags: [Colmeias]
      summary: Send a message to a hive
      description: |
        Dispatches a goal to the hive asynchronously. Before execution, the goal
        is enriched with two layers of memory:

        1. **DB history** — the last 3 completed dispatches for this hive are prepended
           as conversation context so agents maintain continuity across interactions.
        2. **Semantic memory** — if Honeycomb (vector DB) is enabled, the goal is
           embedded and the top-3 most semantically similar past workflow results
           (score > 0.7) are appended as additional context.

        Each hive has its own isolated vector memory collection (keyed by hive ID),
        so memory never leaks between hives.

        - `queen_managed: true` → Queen assembles a dynamic pipeline for this goal.
        - `queen_managed: false` → Uses the hive's pre-defined agents as the pipeline.

        Progress is streamed via WebSocket. Returns `400` if `queen_managed=false`
        and no agents have been defined yet.
      parameters:
        - $ref: '#/components/parameters/ColmeiaID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [goal]
              properties:
                goal:
                  type: string
                  example: "Summarise the latest AI news"
      responses:
        '202':
          description: Mission dispatched asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColmeiaDispatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/colmeias/{id}/historico:
    get:
      tags: [Colmeias]
      summary: List dispatch history
      description: Returns all dispatches for the hive, ordered from newest to oldest.
      parameters:
        - $ref: '#/components/parameters/ColmeiaID'
      responses:
        '200':
          description: Dispatch history
          content:
            application/json:
              schema:
                type: object
                properties:
                  historico:
                    type: array
                    items:
                      $ref: '#/components/schemas/HistoricoDespacho'
                  total:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/colmeias/{id}/agentes:
    parameters:
      - $ref: '#/components/parameters/ColmeiaID'

    get:
      tags: [Colmeias]
      summary: List hive agents
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                type: object
                properties:
                  agentes:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgenteColmeia'
                  total:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Colmeias]
      summary: Add an agent to a hive
      description: |
        Creates a persistent agent inside the hive. The agent's `system_prompt`
        and `allowed_tools` are fully editable via `PUT /api/colmeias/{id}/agentes/{agentId}`.

        **Restriction:** returns `409 Conflict` if the hive has `queen_managed=true`.
        Queen-managed hives assemble agents automatically on every dispatch — pre-defined
        agents are not allowed. Set `queen_managed=false` when creating the hive to use
        custom agents.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertAgenteRequest'
            example:
              name: "Code Reviewer"
              system_prompt: "You are an expert Go code reviewer. Read files and provide detailed feedback."
              allowed_tools: ["read_file", "list_directory", "web_search"]
      responses:
        '201':
          description: Agent added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  agente:
                    $ref: '#/components/schemas/AgenteColmeia'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Hive is queen_managed — pre-defined agents not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Colmeia gerenciada pela Rainha não aceita agentes pré-definidos. Defina queen_managed=false para usar agentes customizados."
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/skills:
    get:
      tags: [Skills]
      summary: List all skills
      responses:
        '200':
          description: Skill list
          content:
            application/json:
              schema:
                type: object
                properties:
                  skills:
                    type: array
                    items:
                      $ref: '#/components/schemas/Skill'
                  total:
                    type: integer
        '428':
          $ref: '#/components/responses/NotConfigured'

    post:
      tags: [Skills]
      summary: Create a skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertSkillRequest'
            example:
              name: "Web Research"
              description: "Pesquisa na web usando DuckDuckGo"
              instructions: "Use web_search para coletar informações atualizadas antes de responder."
              allowed_tools: ["web_search"]
      responses:
        '201':
          description: Skill created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  skill:
                    $ref: '#/components/schemas/Skill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/skills/{id}:
    parameters:
      - $ref: '#/components/parameters/SkillID'

    get:
      tags: [Skills]
      summary: Get a skill
      responses:
        '200':
          description: Skill record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'

    put:
      tags: [Skills]
      summary: Update a skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertSkillRequest'
      responses:
        '200':
          description: Skill updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  skill:
                    $ref: '#/components/schemas/Skill'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Skills]
      summary: Delete a skill
      description: Removes the skill and its associations from all hives and agents.
      responses:
        '200':
          description: Skill deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/colmeias/{id}/skills:
    parameters:
      - $ref: '#/components/parameters/ColmeiaID'

    get:
      tags: [Skills]
      summary: List skills attached to a hive
      responses:
        '200':
          description: Skill list
          content:
            application/json:
              schema:
                type: object
                properties:
                  skills:
                    type: array
                    items:
                      $ref: '#/components/schemas/Skill'
                  total:
                    type: integer
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Skills]
      summary: Attach a skill to a hive
      description: |
        Attaches an existing skill to the hive. On the next `queen_managed` dispatch, the Queen
        receives a `SKILLS DISPONÍVEIS` block in its meta-planning prompt and can assign skills
        to each specialist it creates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skill_id]
              properties:
                skill_id:
                  type: integer
                  example: 1
      responses:
        '200':
          description: Skill attached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '428':
          $ref: '#/components/responses/NotConfigured'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    SetupRequest:
      type: object
      description: Hive configuration payload
      properties:
        api_key:
          type: string
          description: OpenAI API key. Stored securely in the local vault.
          example: "sk-..."
        language:
          type: string
          description: Preferred response language (e.g. "pt-BR", "en-US")
          example: "en-US"
        model:
          type: string
          description: LLM model identifier. Defaults to `gpt-4o-mini`.
          example: "gpt-4o"
          default: "gpt-4o-mini"
        swarm_name:
          type: string
          description: Identifier for the default swarm group. Defaults to `enxame-alfa`.
          example: "my-swarm"
          default: "enxame-alfa"
        max_nectar:
          type: integer
          description: Maximum token budget per workflow. Defaults to `20000`.
          example: 40000
          default: 20000
        max_agents:
          type: integer
          description: Maximum number of concurrent agents.
          example: 5
        supervised:
          type: boolean
          description: When `true`, tool calls require human approval via WebSocket before execution.
          default: false
        isolated:
          type: boolean
          description: When `true`, agents run in isolation mode without shared context.
          default: false
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...